+++ tests/testthat/test-col-types.R
@@ -52,44 +52,54 @@ test_that("inappropriate col_types generate warning", {
   expect_warning(
     read_excel(test_sheet("iris-excel.xlsx"),
                col_types = c("numeric", "text", "numeric", "numeric", "numeric")),
-    "expecting numeric"
+    "Expecting numeric"
   )
   expect_warning(
     read_excel(test_sheet("iris-excel.xls"),
                col_types = c("numeric", "text", "numeric", "numeric", "numeric")),
-    "expecting numeric"
+    "Expecting numeric"
   )
 })
 
-test_that("types imputed & read correctly [xlsx]", {
-  types <- read_excel(test_sheet("types.xlsx"))
-  expect_is(types$number, "numeric")
-  expect_is(types$string, "character")
-  expect_is(types$boolean, "numeric")
+test_that("types guessed correctly [xlsx]", {
+  types <- read_excel(test_sheet("types.xlsx"), sheet = "guess_me")
+  expect_is(types$X__1, "logical")
+  expect_is(types$blank, "logical")
+  expect_is(types$boolean, "logical")
   expect_is(types$date, "POSIXct")
-  expect_is(types$string_in_row_3, "character")
-  skip("switch expectation to logical (vs numeric) when possible (for xls too!)")
+  expect_is(types$numeric, "numeric")
+  expect_is(types$text, "character")
+  expect_true(all(vapply(types, function(x) is.na(x[3]), logical(1))))
 })
 
-test_that("types imputed & read correctly [xls]", {
-  types <- read_excel(test_sheet("types.xls"))
-  expect_is(types$number, "numeric")
-  expect_is(types$string, "character")
-  expect_is(types$boolean, "numeric")
-  #expect_is(types$date, "POSIXct")
-  expect_is(types$string_in_row_3, "character")
-  skip("activate date expectation when xls formula dates are sorted")
+test_that("types guessed correctly [xls]", {
+  types <- read_excel(test_sheet("types.xls"), sheet = "guess_me")
+  expect_is(types$X__1, "logical")
+  expect_is(types$blank, "logical")
+  expect_is(types$boolean, "logical")
+  expect_is(types$date, "POSIXct")
+  expect_is(types$numeric, "numeric")
+  expect_is(types$text, "character")
+  expect_true(all(vapply(types, function(x) is.na(x[3]), logical(1))))
 })
 
 test_that("guess_max is honored for col_types", {
   expect_warning(
-    types <- read_excel(test_sheet("types.xlsx"), guess_max = 2),
-    "expecting numeric"
+    types <- read_excel(
+      test_sheet("types.xlsx"),
+      sheet = "guess_max",
+      guess_max = 2
+    ),
+    "Expecting numeric"
   )
   expect_identical(types$string_in_row_3, c(1, 2, NA))
   expect_warning(
-    types <- read_excel(test_sheet("types.xls"), guess_max = 2),
-    "expecting numeric"
+    types <- read_excel(
+      test_sheet("types.xls"),
+      sheet = "guess_max",
+      guess_max = 2
+    ),
+    "Expecting numeric"
   )
   expect_identical(types$string_in_row_3, c(1, 2, NA))
 })
@@ -132,4 +142,78 @@ test_that("list column reads data correctly [xls]", {
 test_that("setting `na` works in list columns [xls]", {
   na_defined <-  read_excel(test_sheet("list_type.xls"), col_types = "list", na = "a")
   expect_equal(na_defined$var1[[3]], NA)
+})
+
+test_that("contaminated, explicit logical is read as logical", {
+  ## xls
+  expect_warning(
+    df <- read_excel(test_sheet("types.xls"), sheet = "logical_coercion",
+                     col_types = "logical"),
+    "Expecting logical"
+  )
+  expect_is(df$logical, "logical")
+  expect_false(anyNA(df$logical[c(1, 2, 4, 5, 7)]))
+
+  ## xlsx
+  expect_warning(
+    df <- read_excel(test_sheet("types.xlsx"), sheet = "logical_coercion",
+                     col_types = "logical"),
+    "Expecting logical"
+  )
+  expect_is(df$logical, "logical")
+  expect_false(anyNA(df$logical[c(1, 2, 4, 5, 7)]))
+})
+
+test_that("contaminated, explicit date is read as date", {
+  ## xls
+  expect_warning(
+    df <- read_excel(test_sheet("types.xls"), sheet = "date_coercion",
+                     col_types = "date"),
+    "Expecting date"
+  )
+  expect_is(df$date, "POSIXct")
+  expect_false(anyNA(df$date[c(1, 6, 7)]))
+
+  ## xlsx
+  expect_warning(
+    df <- read_excel(test_sheet("types.xlsx"), sheet = "date_coercion",
+                     col_types = "date"),
+    "Expecting date"
+  )
+  expect_is(df$date, "POSIXct")
+  expect_false(anyNA(df$date[c(1, 6, 7)]))
+})
+
+test_that("contaminated, explicit numeric is read as numeric", {
+  ## xls
+  expect_warning(
+    df <- read_excel(test_sheet("types.xls"), sheet = "numeric_coercion",
+                     col_types = "numeric"),
+    "Expecting numeric|Coercing boolean"
+  )
+  expect_is(df$numeric, "numeric")
+  expect_false(anyNA(df$numeric[c(1, 2, 4, 7)]))
+
+  ## xlsx
+  expect_warning(
+    df <- read_excel(test_sheet("types.xlsx"), sheet = "numeric_coercion",
+                     col_types = "numeric"),
+    "Expecting numeric|Coercing boolean"
+  )
+  expect_is(df$numeric, "numeric")
+  expect_false(anyNA(df$numeric[c(1, 2, 4, 7)]))
+})
+
+test_that("contaminated, explicit text is read as text", {
+  ## xls
+  df <- read_excel(test_sheet("types.xls"), sheet = "text_coercion",
+                   col_types = "text")
+  expect_is(df$text, "character")
+  expect_false(anyNA(df$text[-2]))
+
+  ## xlsx
+  df <- read_excel(test_sheet("types.xlsx"), sheet = "text_coercion",
+                   col_types = "text")
+  expect_is(df$text, "character")
+  expect_false(anyNA(df$text[-2]))
 })