+++ NEWS.md
@@ -1,6 +1,8 @@
 # readxl 0.1.1.9000
 
-* Boolean cells are now detected in xls. Suppresses message `"Unknown type: 517"`. For now, Boolean cells in both xls and xlsx become numeric, with values 0 and 1. (#274, #259 @jennybc)
+* `"logical"` is a new column type. When `col_types = NULL`, it is the guessed column type for Boolean cells. It can be specified explicitly in `col_types`. (#?? @jennybc)
+
+* BoolErr cells are now handled in xls. Suppresses message `"Unknown type: 517"`. (#274, #259 @jennybc)
 
 * xls files written by some third party software report both row and column dimensions as 0-indexed, which prevents libxls from reading the last column. Small change to libxls restores access to those cells. (#273, #180, #152, #99 @jennybc)